Add modeline to first line of file to prevent it getting broken by other people
I find that other people editing my code aren't familiar with modelines
and will append code after it if it's at the end of the file, thereby
breaking it. Hopefully by moving it to the beginning of a file it will
be less likely to get broken by other people.